refactor: extract constant_lookup class and define resolver once
